Update main.js
I think the main point of the project (as title says) is to show the Active Forks.
However, when we enter the site and type repo-url ,   it doesn't give the results sorted by the "latest push"  (which is the sign of "activeness"). So there should be `Last Push` instead of  `Stars`. If user wants it to be sorted by *stars*, then s/he should do it explicitly, as the project is not title "most popular forks", instead the project's first action should be as expected by its name "active" forks.

Would be nice if you defaulted that and to meet expectations of users.